test(fdroid): unit-test LibreTranslate code mapping in isolation
